[examples][lg]: Add TableSchema example
